+++ chain-connect/src/types/galaChain.ts
